% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ae_aeacn_ds_disctx_covid.R
\name{check_ae_aeacn_ds_disctx_covid}
\alias{check_ae_aeacn_ds_disctx_covid}
\title{Check if COVID-19 AE indicated DRUG WITHDRAWN then Treatment Discon
indicates ADVERSE EVENT}
\usage{
check_ae_aeacn_ds_disctx_covid(AE, DS, covid_df = NULL)
}
\arguments{
\item{AE}{Adverse Events SDTM dataset with variables USUBJID,AETERM,AEDECOD,AESTDTC,AEACNx}

\item{DS}{Disposition SDTM dataset with variables USUBJID,DSSPID,DSCAT,DSDECOD}

\item{covid_df}{Dataframe of AE terms identify covid, contains variable REFTERM}
}
\value{
boolean value if check returns 0 obs, otherwise return subset dataframe.
}
\description{
This code checks if a patient has COVID-19 AE with AEACN=DRUG WITHDRAWN
then there should also be a corresponding DS.DSDECOD with ADVERSE EVENT.
}
\examples{

covid_df = data.frame(REFTERM = c("COVID-19",
                                  "CORONAVIRUS POSITIVE"
                                  )
                     )
AE <- data.frame(
    STUDYID = 1,
    USUBJID = c(1,2,3,1,2,3),
    AESTDTC = '2020-05-05',
    AETERM  = c("abc Covid-19", "covid TEST POSITIVE",rep("other AE",4)),
    AEDECOD = c("COVID-19", "CORONAVIRUS POSITIVE", rep("OTHER AE",4)),
    AEACN = c("DRUG WITHDRAWN", rep("DOSE NOT CHANGED",5)),
    stringsAsFactors = FALSE
)

DS <- data.frame(
 USUBJID = c(1,1,2,3,4),
 DSSPID  = 'XXX-DISCTX-XXX',
 DSCAT   = "DISPOSITION EVENT",
 DSDECOD = "REASON",
 stringsAsFactors = FALSE
)

check_ae_aeacn_ds_disctx_covid(AE, DS, covid_df)

DS[1, "DSDECOD"] <- 'ADVERSE EVENT'
check_ae_aeacn_ds_disctx_covid(AE, DS, covid_df)

}
\author{
Sarwan Singh
}
