% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ae_aesdth_aedthdtc.R
\name{check_ae_aesdth_aedthdtc}
\alias{check_ae_aesdth_aedthdtc}
\title{Check AEs with AESDTH of "Y" but No AEDTHDTC Value}
\usage{
check_ae_aesdth_aedthdtc(AE, preproc = identity, ...)
}
\arguments{
\item{AE}{Adverse Event SDTM dataset with variables USUBJID, AEDTHDTC,
AESDTH, AEDECOD and AESTDTC}

\item{preproc}{An optional company specific preprocessing script}

\item{...}{Other arguments passed to methods}
}
\value{
Boolean value for whether the check passed or failed, with 'msg'
attribute if the check failed
}
\description{
This check looks for AE entries with AESDTH of "Y"
but no AEDTHDTC (death date) value
}
\examples{

AE <- data.frame(USUBJID = c(1:5), AEDTHDTC = c(1:2, NA, "NA", 3),
                 AESDTH = c("", rep("Y",2), rep("", 2)),
                 AEDECOD = letters[1:5], AESTDTC = c(1:5),
                 AESPID = "FORMNAME-R:5/L:5XXXX",
                 stringsAsFactors=FALSE)

check_ae_aesdth_aedthdtc(AE)
check_ae_aesdth_aedthdtc(AE,preproc=roche_derive_rave_row)

}
\author{
Shumei Chi
}
