% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ae_fatal.R
\name{check_ae_fatal}
\alias{check_ae_fatal}
\title{Check for death variable consistency when AEOUT=="FATAL"}
\usage{
check_ae_fatal(AE, preproc = identity, ...)
}
\arguments{
\item{AE}{Adverse Events SDTM dataset with variables USUBJID, AEDECOD,
AESTDTC, AEDTHDTC, AEOUT, AESDTH}

\item{preproc}{An optional company specific preprocessing script}

\item{...}{Other arguments passed to methods}
}
\value{
boolean value if check failed or passed with 'msg' attribute if the
test failed
}
\description{
This check looks for consistency in AESDTH, AEDTHDTC, and
AETOXGR (if applicable) when AEOUT is 'FATAL'
}
\examples{

AE <- data.frame(
 STUDYID = "Study1",
 USUBJID = 1:5,
 AESTDTC = "01JAN2017",
 AEDECOD = c("AE1","AE2","AE3","AE4","AE5"),
 AEOUT = "FATAL",
 AEDTHDTC = c("01FEB2017",NA,"02FEB2017","03FEB2017",NA),
 AESDTH = c("Y","Y","N","Y",NA),
 AETOXGR = c("5","5","5",NA,NA),
 AESPID = rep("FORMNAME-R:13/L:13XXXX",5),
 OTHERVAR = NA,
 stringsAsFactors = FALSE
)

check_ae_fatal(AE)
check_ae_fatal(AE,preproc=roche_derive_rave_row)

AE$AETOXGR <- NULL
check_ae_fatal(AE)

AE$AEDECOD <- NULL
check_ae_fatal(AE)

}
\author{
Aldrich Salva
}
