% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_datasets_for_testing.R
\name{check_dummy_datasets}
\alias{check_dummy_datasets}
\title{Puts into environment a list of dummy datasets for testing.}
\usage{
check_dummy_datasets()
}
\value{
Nested list of dataframes.
}
\description{
Ad-hoc named list of dummy dataframes first used for testing each check function and copied here.  List is structured first by name of check function and then 0, 1 or more dataframes.  Before use, the list must be unnested and assembled as required.
}
\examples{
dummy_datasets = check_toy_datasets()

L = check_dummy_datasets()
identical(L[[c("check_ae_aeacnoth", "pass","one")]], L$check_ae_aeacnoth$pass$one)
}
\author{
jim rothstein \link{14JUNE2024}
}
