% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ex_exstdtc_after_rs_rsdtc.R
\name{check_ex_exstdtc_after_rs_rsdtc}
\alias{check_ex_exstdtc_after_rs_rsdtc}
\title{Check EXSTDTC records occurring on or after RSDTC}
\usage{
check_ex_exstdtc_after_rs_rsdtc(EX, RS, preproc = identity, ...)
}
\arguments{
\item{EX}{Exposure SDTM dataset with variables USUBJID, EXSTDTC, and EXDOSE}

\item{RS}{Disease Response SDTM dataset with variables USUBJID, RSDTC,
RSTESTCD, and RSEVAL}

\item{preproc}{An optional company specific preprocessing script}
}
\value{
boolean value if check failed or passed with 'msg' attribute if the
test failed
}
\description{
This check identifies records where EXSTDTC dates occurs
on or after RSDTC. Only applies to assessments by investigator,
RSEVAL = "INVESTIGATOR",RSTESTCD == "OVRLRESP" and where EXDOSE is greater than 0.
}
\examples{


# No errors
EX_SAMPLE <- data.frame(USUBJID = LETTERS[1:5],
                 EXSTDTC = c("2015-12-31","2016-12-31","2017-12-31","2009-12-31","2001-12-31" ),
                 EXDOSE = 1:5,
                 stringsAsFactors = FALSE)


RS_SAMPLE<- data.frame(USUBJID = LETTERS[1:5],
                RSDTC=c("2016-02-01T08:25", "2018-01-05T09:25",
                       "2022-03-15T10:25","2017-01-20T08:25","2017-01-25T08:25",
                       "2017-02-01T08:25", "2019-01-05T09:25",
                       "2023-03-15T10:25","2018-01-20T08:25","2018-01-25T08:25"),
                RSTESTCD="OVRLRESP",
                RSEVAL="INVESTIGATOR",
                RSSPID = "FORMNAME-R:19/L:19XXXX",
                stringsAsFactors=FALSE)

check_ex_exstdtc_after_rs_rsdtc(EX= EX_SAMPLE,RS= RS_SAMPLE)



# Four errors, preproc=identity
EX_SAMPLE <- data.frame(USUBJID = LETTERS[1:5],
                 EXSTDTC = c("2015-12-31","2016-12-31","2017-12-31","2009-12-31","2001-12-31" ),
                 EXDOSE = 1:5,
                 stringsAsFactors = FALSE)


RS_SAMPLE<- data.frame(USUBJID = LETTERS[1:5],
                RSDTC=c("2015-12-31T08:25", "2018-01-05T09:25",
                       "2017-03-15T10:25","2017-01-20T08:25","2017-01-25T08:25",
                       "1998-02-01T08:25", "2019-01-05T09:25",
                       "2016-03-15T10:25","2018-01-20T08:25","2018-01-25T08:25"),
                RSTESTCD="OVRLRESP",
                RSEVAL="INVESTIGATOR",
                RSSPID = "FORMNAME-R:19/L:19XXXX",
                stringsAsFactors=FALSE)

check_ex_exstdtc_after_rs_rsdtc(EX= EX_SAMPLE,RS= RS_SAMPLE)



# missing column
EX_SAMPLE <- data.frame(USUBJID = LETTERS[1:5],
                 # EXSTDTC = c("2015-12-31","2016-12-31","2017-12-31","2009-12-31","2001-12-31" ),
                 EXDOSE = 1:5,
                 stringsAsFactors = FALSE)


RS_SAMPLE<- data.frame(USUBJID = LETTERS[1:5],
                RSDTC=c("1999-02-01T08:25", "2018-01-05T09:25",
                       "2017-03-15T10:25","2017-01-20T08:25","2017-01-25T08:25",
                       "1998-02-01T08:25", "2019-01-05T09:25",
                       "2016-03-15T10:25","2018-01-20T08:25","2018-01-25T08:25"),
                RSTESTCD="OVRLRESP",
                RSEVAL="INVESTIGATOR",
                RSSPID = "FORMNAME-R:19/L:19XXXX",
                stringsAsFactors=FALSE)

check_ex_exstdtc_after_rs_rsdtc(EX= EX_SAMPLE,RS= RS_SAMPLE)




# Return df with errors
EX_SAMPLE <- data.frame(USUBJID = LETTERS[1:5],
                 EXSTDTC = c("2015-12-31","2016-12-31","2017-12-31","2009-12-31","2001-12-31" ),
                 EXDOSE = 1:5,
                 stringsAsFactors = FALSE)


RS_SAMPLE<- data.frame(USUBJID = LETTERS[1:5],
                RSDTC=c("2015-12-31T08:25", "2018-01-05T09:25",
                       "2017-03-15T10:25","2017-01-20T08:25","2017-01-25T08:25",
                       "1998-02-01T08:25", "2019-01-05T09:25",
                       "2016-03-15T10:25","2018-01-20T08:25","2018-01-25T08:25"),
                RSTESTCD="OVRLRESP",
                RSEVAL="INVESTIGATOR",
                RSSPID = "FORMNAME-R:19/L:19XXXX",
                stringsAsFactors=FALSE)

check_ex_exstdtc_after_rs_rsdtc(EX= EX_SAMPLE,RS= RS_SAMPLE, preproc=roche_derive_rave_row)
}
\author{
Josue Alcaraz
}
