% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sc_dm_seyeselc.R
\name{check_sc_dm_seyeselc}
\alias{check_sc_dm_seyeselc}
\title{Check if Subject Characteristics, Study Eye Selection data is missing}
\usage{
check_sc_dm_seyeselc(DM, SC)
}
\arguments{
\item{DM}{Subject Demographics Dataset with variable USUBJID}

\item{SC}{Subject Characteristics Dataset for Ophtho Study with variables USUBJID, SCTESTCD, SCTEST, SCCAT, SCORRES, SCDTC}
}
\description{
Check If SC.SCCAT = "STUDY EYE SELECTION" and SCTESTCD = "FOCID", Check if SC.SCORRES has OS, OD, or OU. Flag if Missing or any other value.
}
\examples{

dm <- data.frame(USUBJID = c(1,2))
sc <- data.frame(USUBJID  = c(1,1,1,2,2,2),
                 SCTEST   = c("Eye Meeting Eligibility Criteria",
                              "Focus of Study-Specific Interest",
                              " ",
                              "Eye Meeting Eligibility Criteria",
                              "Focus of Study-Specific Interest", " "),
                 SCTESTCD = c("ELIGEYE", "FOCID", "", "ELIGEYE", "FOCID", ""),
                 SCCAT    = c("STUDY EYE SELECTION",
                              "STUDY EYE SELECTION",
                              "",
                              "STUDY EYE SELECTION",
                              "STUDY EYE SELECTION", ""),
                 SCORRES  = c("LEFT", "OS", "", "RIGHT", "OD", ""),
                 SCDTC    = rep("2021-01-01", 6),
                 stringsAsFactors = FALSE)

check_sc_dm_seyeselc(SC=sc, DM=dm)

dm <- data.frame(USUBJID = c(1,2,3,4))
sc <- data.frame(USUBJID  = c(1,1,1,2,2,2),
                 SCTEST   = c("Eye Meeting Eligibility Criteria",
                              "Focus of Study-Specific Interest",
                              " ",
                              "Eye Meeting Eligibility Criteria",
                              "Focus of Study-Specific Interest",
                              " "),
                 SCTESTCD = c("ELIGEYE", "FOCID", "", "ELIGEYE", "FOCID", ""),
                 SCCAT    = c("STUDY EYE SELECTION",
                              "STUDY EYE SELECTION",
                              "",
                              "STUDY EYE SELECTION",
                              "STUDY EYE SELECTION", ""),
                 SCORRES  = c("LEFT", "OS", "", "RIGHT", "", ""),
                 SCDTC    = rep("2021-01-01", 6),
                 stringsAsFactors = FALSE)

check_sc_dm_seyeselc(SC=sc, DM=dm)


}
\author{
Monarch Shah (HackR 2021 Team Eye)
}
