% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_toy_datasets}
\alias{check_toy_datasets}
\title{Puts into environment a list of toy datasets for testing.}
\usage{
check_toy_datasets()
}
\value{
Nested list of dataframes.
}
\description{
Ad-hoc named list of toy dataframes first used for testing each check function and copied here.  List is structured first by name of check function and then 0, 1 or more dataframes.  Before use, the list must be unnested and assembled as required.
}
\examples{
toy_datasets = check_toy_datasets()

}
\author{
jim rothstein \link{14JUNE2024}
}
