% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tr_trdtc_across_visit.R
\name{check_tr_trdtc_across_visit}
\alias{check_tr_trdtc_across_visit}
\title{Check TR Longest Diameter Records where the same date occurs accross multiple visits}
\usage{
check_tr_trdtc_across_visit(TR)
}
\arguments{
\item{TR}{Tumor Result SDTM dataset with variables USUBJID, TRDTC, VISIT, TRTESTCD}
}
\value{
boolean value if check failed or passed with 'msg' attribute if the
test failed
}
\description{
This check identifies records where the same date TRDTC occurs
across multiple visits when TRTESTCD is "LDIAM". Only applies to assessments by investigator.
}
\examples{

TR <- data.frame(
USUBJID = 1,
TRDTC = c(rep("2016-01-01",3), rep("2016-06-01",5), rep("2016-06-24",2)),
VISIT = c(rep("C1D1",3), rep("C1D2",3), rep("C2D1",4)),
TRTESTCD = c(rep("LDIAM",7),rep("SAXIS",3)),
stringsAsFactors=FALSE)

check_tr_trdtc_across_visit(TR)


}
\author{
Will Harris
}
