% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tu_rs_new_lesions.R
\name{check_tu_rs_new_lesions}
\alias{check_tu_rs_new_lesions}
\title{Check for consistency between new lesions and overall PD response}
\usage{
check_tu_rs_new_lesions(RS, TU)
}
\arguments{
\item{RS}{Response SDTM dataset with variables USUBJID, RSSTRESC, RSTESTCD}

\item{TU}{Tumor Identification SDTM dataset with variables USUBJID, TUSTRESC,
TUDTC}
}
\value{
TRUE if check is passed and FALSE if check failed + 'msg' and 'data'
attributes
}
\description{
This checks for patients with new lesions in TU (TUSTRESC='NEW')
but no overall response of PD in RS. Only applies to assessments by
investigator.
}
\examples{

TU <- data.frame(
 USUBJID = 1:3,
 TUSTRESC = c("INV001","NEW","NEW"),
 TUDTC = "2017-01-01"
)

RS <- data.frame(
 USUBJID = 1:2,
 RSTESTCD = "OVRLRESP",
 RSSTRESC = c("SD","NE")
)

check_tu_rs_new_lesions(RS,TU)


}
\author{
Will Harris
}
