% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_R_script}
\alias{create_R_script}
\title{Create .R file with sdtmchecks function calls}
\usage{
create_R_script(metads = sdtmchecksmeta, file = "sdtmchecks_run_all.R")
}
\arguments{
\item{metads}{sdtmchecksmeta file}

\item{file}{filename and/or path to save to}
}
\value{
R script with user specified sdtmchecks based on sdtmchecksmeta file
}
\description{
Function that uses sdtmchecksmeta as input and creates .R file with function calls
}
\examples{

\dontrun{

create_R_script(file = "run_the_checks.R")

# Only include selected checks
mymetads = sdtmchecksmeta \%>\% 
filter(category == "ALL" & priority == "High")

create_R_script(metads = mymetads, file = "run_the_checks.R")

#Roche specific function calls
mymetads = sdtmchecksmeta \%>\% 
mutate(fxn_in=fxn_in_roche)

create_R_script(metads = mymetads, file = "run_the_checks.R")

}
}
\author{
Monarch Shah
}
