% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diff_reports}
\alias{diff_reports}
\title{Create a sdtmchecks list object with column indicating whether the issue was previously seen}
\usage{
diff_reports(old_report, new_report)
}
\arguments{
\item{old_report}{an older sdtmchecks list object as created by \code{run_all_checks}}

\item{new_report}{a newer sdtmchecks list object as created by \code{run_all_checks}}
}
\value{
list
}
\description{
This function will compare two sdtmchecks list objects as created by \code{run_all_checks}.
It creates a report with only records seen in the more recent report (ie \code{new_report}).
It then adds a flag for whether this record was seen in the old report (ie \code{old_report}).
This function ignores old results not in new report (because presumably they were resolved).
A left join is performed, so it matters which report you define as 'new' and 'old'.
}
\examples{
\dontrun{
  diff_reports(old_report=sdtmchecks_results_15JUL2022,
               new_report=sdtmchecks_results_15AUG2022)
}


}
