% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{domains}
\alias{domains}
\title{Domains for sdtmchecks}
\format{
A data frame with a row for each SDTM domain with at least one check in the package:
\describe{
\item{domains}{Lowercased SDTM domain name}
}
}
\usage{
data(domains)
}
\description{
A dataset containing domains with checks in the SDTM checks package. The data can be used as input into functions.
}
\examples{
data(domains)
head(domains[,1:5])

\dontrun{
 domains
}
}
\keyword{datasets}
