% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dtc_dupl_early}
\alias{dtc_dupl_early}
\title{Check if start dates are duplicated or earlier than prior visit date}
\usage{
dtc_dupl_early(dts, vars, groupby, dtc, ...)
}
\arguments{
\item{dts}{dataset, e.g. EX}

\item{vars}{variables in a form c("USUBJID", "EXTRT", "VISITNUM", "VISIT", "EXSTDTC")}

\item{groupby}{variables used for grouping and visit.order derivation}

\item{dtc}{the date variable}

\item{...}{variables used for ordering before visit.order derivation}
}
\value{
dataframe with records of duplicated or earlier than last visit date
}
\description{
This is a core function for checking if start dates are
duplicated or earlier than last visit's (possible datetime data entry error),
can be used in separate checks for each domain
}
\author{
James Zhang
}
\keyword{internal}
