% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%has_all\%}
\alias{\%has_all\%}
\title{Check if data frame has all specified variables}
\usage{
df \%has_all\% varnames
}
\arguments{
\item{df}{a data frame}

\item{varnames}{a vector with variable names}
}
\value{
boolean
}
\description{
Check if data frame has all specified variables
}
