% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%has_any\%}
\alias{\%has_any\%}
\title{Check if data frame has at least one of the specified variables}
\usage{
df \%has_any\% varnames
}
\arguments{
\item{df}{a data frame}

\item{varnames}{a vector with variable names}
}
\value{
boolean
}
\description{
Check if data frame has at least one of the specified variables
}
