% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%lacks_all\%}
\alias{\%lacks_all\%}
\title{Check if data frame is missing any of the specified variables}
\usage{
df \%lacks_all\% varnames
}
\arguments{
\item{df}{a data frame}

\item{varnames}{a vector with variable names}
}
\value{
boolean
}
\description{
Check if data frame is missing any of the specified variables
}
