% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%lacks_any\%}
\alias{\%lacks_any\%}
\title{Check if data frame has at least one variable from a set of specified variables}
\usage{
df \%lacks_any\% varnames
}
\arguments{
\item{df}{a data frame}

\item{varnames}{a vector with variable names}
}
\value{
boolean
}
\description{
Check if data frame has at least one variable from a set of specified variables
}
