% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{impute_day01}
\alias{impute_day01}
\title{Add day of "01" to dates that are in the format of "yyyy-mm"}
\usage{
impute_day01(dates)
}
\value{
string
}
\description{
Add day of "01" to dates that are in the format of "yyyy-mm"
}
\keyword{internal}
