% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lacks_msg}
\alias{lacks_msg}
\title{Check data frame for mandatory variables and specify which are missing in
message}
\usage{
lacks_msg(df, varnames)
}
\arguments{
\item{df}{a data frame}

\item{varnames}{a vector with variable names}
}
\value{
character string saying which variables are missing
}
\description{
Check data frame for mandatory variables and specify which are missing in
message
}
