% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{missing_month}
\alias{missing_month}
\title{Function to check if month is missing while year and day are non-missing
(i.e. would be in the format of "yyyy---dd")}
\usage{
missing_month(date)
}
\arguments{
\item{date}{date vector (character) in the format 2020-01-20}
}
\value{
vector
}
\description{
Function to check if month is missing while year and day are non-missing
(i.e. would be in the format of "yyyy---dd")
}
\keyword{internal}
