% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nickname}
\alias{nickname}
\title{Nickname for package version}
\format{
An object of class \code{character} of length 1.
}
\usage{
nickname
}
\description{
A nickname associated with the package version.  Previously lived in the description file.
}
\keyword{internal}
