% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{report_to_xlsx}
\alias{report_to_xlsx}
\title{Save report as an xlsx file}
\usage{
report_to_xlsx(res, outfile, extrastring = "")
}
\arguments{
\item{res}{results list created by run_all_checks}

\item{outfile}{file path/name to write to}

\item{extrastring}{optionally display extra info alongside version info, e.g. diff info}
}
\value{
xlsx file
}
\description{
Save report as an xlsx file
}
\examples{

\donttest{

ae = haven::read_sas("path/to/ae.sas7bdat")
cm = haven::read_sas("path/to/cm.sas7bdat")
dm = haven::read_sas("path/to/dm.sas7bdat")

all_rec=run_all_checks(verbose = TRUE)

report_to_xlsx(res=all_rec,outfile="check_results.xlsx")


}
}
\seealso{
Example programs for running data checks
\code{\link{diff_reports}()},
\code{\link{run_all_checks}()},
\code{\link{run_check}()}
}
\concept{ex_rpt}
\keyword{ex_rpt}
\keyword{internal}
