% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtmchecks-package.R
\docType{package}
\name{sdtmchecks-package}
\alias{sdtmchecks}
\alias{sdtmchecks-package}
\title{sdtmchecks: Data Quality Checks for SDTM Datasets}
\description{
Code to identify common data issues in SDTM data. These checks are intended to be generalizable, actionable, and meaningful for analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/sdtmchecks/}
  \item \url{https://github.com/pharmaverse/sdtmchecks}
  \item Report bugs at \url{https://github.com/pharmaverse/sdtmchecks/issues}
}

}
\author{
\strong{Maintainer}: Will Harris \email{harric17@gene.com} (\href{https://orcid.org/0000-0002-9807-1274}{ORCID})

}
\keyword{internal}
