% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{sdtmchecks}
\alias{sdtmchecks}
\title{sdtmchecks: A package containing checks for common SDTM issues}
\description{
Package containing checks for common SDTM issues.
This package contains functions to identify common data issues in SDTM data.
These checks are intended to be generalizable, actionable, and meaningful for analysis.
}
