% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{sdtmchecksSearch}
\alias{sdtmchecksSearch}
\title{Search data check functions available in sdtmchecksmeta}
\usage{
sdtmchecksSearch()
}
\description{
Interactively search the check functions of the sdtmchecks package
}
\details{
This addin can be used to interactively subset the dataframe
}
