% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xlsx2list}
\alias{xlsx2list}
\title{Create list from tabs of spreadsheet (*xlsx)}
\usage{
xlsx2list(rptwb, firstrow = 1, verbose = TRUE)
}
\arguments{
\item{rptwb}{List of check results in newer report read in via openxlsx::loadWorkbook}

\item{firstrow}{Row input for startRow parameter of openxlsx::read.xlsx, default is 1}

\item{verbose}{Print information to console}
}
\value{
list named xlsxfile
}
\description{
This creates a list based on the sdtmchecks_yyyy-mm-dd.xlsx file, such as report created via \code{report_to_xlsx()}
}
\examples{

\dontrun{

rptfile <- "Downloads/sdtmchecks_GO12345_2021-03-12.xlsx"
rptwb <- openxlsx::loadWorkbook(rptfile)
names(rptwb)
rptwb

xlsx2list(rptwb=rptwb, firstrow=1)

}
}
\seealso{
Reporting-related utility functions
\code{\link{convert_var_to_ascii}()},
\code{\link{create_R_script}()},
\code{\link{truncate_var_strings}()}
}
\concept{utils_rpt}
\keyword{utils_rpt}
