% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rs_rsdtc_across_visit.R
\name{check_rs_rsdtc_across_visit}
\alias{check_rs_rsdtc_across_visit}
\title{Check RS Records where the same date occurs across multiple visits}
\usage{
check_rs_rsdtc_across_visit(RS)
}
\arguments{
\item{RS}{Disease Response SDTM dataset with variables USUBJID, RSDTC, VISIT}
}
\value{
boolean value if check failed or passed with 'msg' attribute if the
test failed
}
\description{
This check identifies records where the same date RSDTC occurs
across multiple visits. Only applies to assessments by investigator.
}
\examples{

RS <- data.frame(
USUBJID = 1,
RSDTC = c(rep("2016-01-01",3), rep("2016-06-01",5), rep("2016-06-24",2)),
VISIT = c(rep("C1D1",3), rep("C1D2",3), rep("C2D1",4)),
stringsAsFactors=FALSE)

check_rs_rsdtc_across_visit(RS)



}
\author{
Will Harris
}
