% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tr_dup.R
\name{check_tr_dup}
\alias{check_tr_dup}
\title{Check for duplicate TR records}
\usage{
check_tr_dup(TR)
}
\arguments{
\item{TR}{TR dataframe with variables USUBJID, TRLINKID, TRTESTCD, TRCAT, TRSTRESC,
TRSTRESU, VISITDY, TRDTC}
}
\description{
This check looks for duplicate TR records and returns a data frame.
Only applies to assessments by investigator.
}
\examples{

 TR <- data.frame(
 USUBJID  = c(1,1,2,2),
 TRCAT    = c(1,1,2,2),
 TRTESTCD = c(1,1,2,2),
 TRLINKID = c(1,1,2,2),
 TRSPID   = c(1,1,2,2),
 TRDTC    = c(1,1,2,2),
 TRSTRESC = c(1,1,2,2)
)

check_tr_dup(TR)


}
\author{
Joel Laxamana
}
