% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diff_reports}
\alias{diff_reports}
\title{Diff two sdtmchecks reports}
\usage{
diff_reports(old_report, new_report)
}
\arguments{
\item{old_report}{older sdtmchecks report as generated by sdtmchecks::run_all_checks()}

\item{new_report}{newer sdtmchecks report as generated by sdtmchecks::run_all_checks()}
}
\value{
list of sdtmchecks results based on new_report with report_diff indicator
}
\description{
This report will identify flagged records from sdtmcheck functions
that are "new" and those that are "old" for a study.
This will help quickly target newly emergent issues that may require a new query or
investigation while indicating issues that were encountered from a prior
report and may have already been queried.

This diff_reports() function requires a newer and older set of results from
sdtmchecks::run_all_checks(), which will generate a list of check results.
A added column "report_diff" is created with values of "NEW" and "PRIOR"
in the list of check results, flagging whether a given record that is present
in the new result is also present in the old result.

It makes a difference which report is defined as "new" and "old".
This code only keeps results flagged in the new report and ignores
old results not in new report (because they were resolved and no longer flagged
as an issue in SDTM).
}
\examples{
\dontrun{

old = readRDS("saved_reports/sdtmchecks_01JAN2023.rds")
new = readRDS("saved_reports/sdtmchecks_01FEB2023.rds")
res=diff_reports(old_report=old,new_report=new)

# Write out result
report_to_xlsx(res,outfile=paste0("saved_reports/sdtmchecks_diff_",Sys.Date(),".xlsx"))

}
}
\seealso{
Example programs for running data checks
\code{\link{report_to_xlsx}()},
\code{\link{run_all_checks}()},
\code{\link{run_check}()}
}
\concept{ex_rpt}
\keyword{ex_rpt}
