% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diff_reports}
\alias{diff_reports}
\title{Diff two sdtmchecks reports}
\usage{
diff_reports(old_report, new_report)
}
\arguments{
\item{old_report}{older sdtmchecks report as generated by run_all_checks}

\item{new_report}{newer sdtmchecks report as generated by run_all_checks}
}
\value{
list of diffed sdtmchecks results
}
\description{
run_all_checks will generate a list of check results.  This function
takes two run_all_checks results, a newer one and an older one, and creates a column
flagging whether a given record that is present in the new result is also present in
the old result.  This might be helpful for quickly identifying new issues.
}
\examples{
\dontrun{

old = readRDS("saved_reports/sdtmchecks_01JAN2023.rds")
new = readRDS("saved_reports/sdtmchecks_01FEB2023.rds")
res=diff_reports(old_report=old,new_report=new)

# Write out result
report_to_xlsx(res,outfile=paste0("saved_reports/sdtmchecks_diff_",Sys.Date(),".xlsx"))

}
}
