% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fail}
\alias{fail}
\title{Fail object}
\usage{
fail(msg, data = NULL)
}
\arguments{
\item{msg}{character strings with fail message}

\item{data}{a data frame to be printed with the fail message}
}
\value{
Boolean with msg and data attributes
}
\description{
With Message and part of the Data that is responsible for the check to fail
}
