% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all_checks.R
\name{run_all_checks}
\alias{run_all_checks}
\title{Run all checks in sdtmchecks package using parallel processing}
\usage{
run_all_checks(
  metads = sdtmchecksmeta,
  priority = c("High", "Medium", "Low"),
  type = c("ALL", "ONC", "COVID", "PRO", "OPHTH"),
  verbose = TRUE,
  ncores = 1,
  run_env = parent.frame()
)
}
\arguments{
\item{metads}{Metadata to use to execute the checks.  The default is the sdtmchecksmeta dataframe available in the packages.  This object could easily be customized, e.g subset, etc.}

\item{priority}{Type of data checks priority, i.e. c("High", "Medium", "Low"). NULL runs all priority.}

\item{type}{Type of data checks, i.e. c("ALL", "ONC", "COV", "PRO", "OPHTH"). NULL runs all type.}

\item{verbose}{Whether to display messages while running}

\item{ncores}{Number of cores for parallel processing, default is 1 (sequential)}

\item{run_env}{environment the function is executed in}
}
\value{
list with results from individual checks
}
\description{
This function runs all checks in the sdtmchecks package.  It expects sdtm domains saved as dataframe objects in your global environment.  These dataframes should have lowercase names, e.g. dm.
}
\details{
to look up documentation for the data checks package, please use command ??sdtmchecks
}
\examples{

\dontrun{

# Assuming sdtm datasets are in your global environment
all_rec<-run_all_checks(metads=sdtmchecksmeta, 
                        priority=c("High","Medium","Low"), 
                        type=c("ALL", "ONC", "COVID", "PRO"))
all_rec<-run_all_checks(metads=sdtmchecksmeta, 
                        priority=NULL, 
                        type=c("ONC", "COVID", "PRO"))
}
}
\seealso{
Example programs for running data checks
\code{\link{report_to_xlsx}()},
\code{\link{run_check}()}
}
\concept{ex_rpt}
\keyword{ex_rpt}
