% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtmchecks-package.R
\docType{package}
\name{sdtmchecks-package}
\alias{sdtmchecks}
\alias{sdtmchecks-package}
\title{sdtmchecks: Data Quality Checks for Study Data Tabulation Model (SDTM) Datasets}
\description{
A series of checks to identify common issues in Study Data Tabulation Model (SDTM) datasets. These checks are intended to be generalizable, actionable, and meaningful for analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/sdtmchecks/}
  \item \url{https://github.com/pharmaverse/sdtmchecks}
  \item Report bugs at \url{https://github.com/pharmaverse/sdtmchecks/issues}
}

}
\author{
\strong{Maintainer}: Will Harris \email{harric17@gene.com}

Authors:
\itemize{
  \item Sara Bodach
  \item Edgar Manukyan
  \item Adrian Waddell
  \item Ross Farrugia
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
