% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{sdtmchecksSearch}
\alias{sdtmchecksSearch}
\title{Search data check functions available in sdtmchecksmeta}
\usage{
sdtmchecksSearch()
}
\description{
Interactively search the check functions of the sdtmchecks package
}
\details{
This RStudio Addin (for RStudio v0.99.878+) can be used to search
and subset the sdtmchecksmeta dataframe, when using R Studio interactively.
The data check functions searched all start with "check_..."
}
\keyword{internal}
